requirements.txt
numpy
python-socketio
aiohttp
requests
boto3
pyyaml
datasets
transformers
opacus
gym
zstd
torch-optimizer
timm
pandas
wandb
torchmetrics